[#188] Update package.json and tsconfig.prod.json
The update to package.json sets the required version of node to 10.10.0. This is the highest version we can support in RHEL 7 environments as it is the latest version offered by Software Collection Libraries. This is a requirement due to the
way we're currently integrating rhd-frontend with Drupal.

The update to tsconfig.prod.json removes the compiler option 'allowUmdGlobalAccess' as this was only introduced in Typescript 3.5 and this project currently does not support that.