Add component-specific code snippet styles
This adds a few styles for the code snippet assembly type that are
specific to the assembly type implementation i.e. not styles that solely
apply to the <pre> or <code> HTML elements. These styles are provided by
HighlightJS, and we need to override them to meet our design
requirements.