[MWES-3660] Support new SSO servers
Two updates in this commit.

Firstly we update drupal-namepsace.js to read the values of the Keycloak client_id and realm from the environment properties provided by Drupal. Secondly we update sso.js to ensure that Keycloak initialisation is updated to work with these configured values rather than being hardcoded.