[MWES-3660] Update to support new SSO servers
This commit updates the rhd-frontend component to expose the new SSO configuration variables provided by Drupal and then use those variables in the configuration of the Keycloak initialisation process. This removes the hard-coding of our current SSO server and make it much easier for us to make the transition to the new ones.